opens postgresql port to the host operating system.
`sonar-runner` and other sonar clients will store their reports directly
in the database. This will only work if the database is publically reachable

defined fixed port forwardings for postgres and sonar. Anything else
will be very inconvenient for using with sonar, as the ports need to
be hardcoded in the configuration of the sonar client.